Tutorial 6 - Property Binding
HTML Attribute and DOM property difference.

Ex. <input type="test" value="Ho Wai">

Gets the value of the current inspected element
$0.getAttribute('value')
"Ho Wai"
$0.value
"Ho Wai"

But if you change the value inside the test box:
$0.getAttribute('value')
"Ho Wai"
$0.value
"Codevolution"

Attributes and properties are different because:

Attributes - HTML
Properties - DOM

Attributes initialize DOm properties. Attributes values cannot change once they are initialized.
Property values can change.

Propery binding is necessary because interpolation has a limitation which is that it only works with string values.

There are HTML properties which are boolean values which need to be binded to as well.

For true:
 <input id="{{myId}}" value="Ho Wai">
For false:
 <input disabled id="{{myId}}" value="Ho Wai">
<input disabled="false" id="{{myId}}" value="Ho Wai">
 <input disabled="{{false}}" id="{{myId}}" value="Ho Wai">
Disables input. But even though we assign the value of false to disabled, it'll still initialize the property.

This is a problem with boolean attributes.

Solution:
<input [disabled]="false" id="{{myId}}" value="Ho Wai">
-Put the attribute into square brackets.
<input [disabled]="false" id="{{myId}}" value="Ho Wai">

-Or define a variable:
public isDisabled = false;
<input [disabled]="isDisabled" id="{{myId}}" value="Ho Wai">
-Alternate:
<input bind-disabled="isDisabled" id="{{myId}}" value="Ho Wai">